mavsdk-aerobridge/requirements.txt
mavsdk==0.22.3
requests==2.27.1
python-dotenv==0.19.2
